use target route query for transaction query
